[feature] 회원탈퇴 버튼 위치 변경
